+++ spring-boot-starters/wx-java-open-spring-boot-starter/src/main/java/com/binarywang/spring/starter/wxjava/open/config/WxOpenAutoConfiguration.java
@@ -12,6 +12,9 @@ import org.springframework.context.annotation.Import;
  */
 @Configuration
 @EnableConfigurationProperties(WxOpenProperties.class)
-@Import({WxOpenStorageAutoConfiguration.class, WxOpenServiceAutoConfiguration.class})
+@Import({
+  WxOpenStorageAutoConfiguration.class,
+  WxOpenServiceAutoConfiguration.class
+})
 public class WxOpenAutoConfiguration {
 }
