+++ spring-boot-starters/wx-java-open-spring-boot-starter/src/main/java/com/binarywang/spring/starter/wxjava/open/config/storage/WxOpenInMemoryConfigStorageConfiguration.java
@@ -1,0 +1,39 @@
+package com.binarywang.spring.starter.wxjava.open.config.storage;
+
+import com.binarywang.spring.starter.wxjava.open.properties.WxOpenProperties;
+import lombok.RequiredArgsConstructor;
+import me.chanjar.weixin.open.api.WxOpenConfigStorage;
+import me.chanjar.weixin.open.api.impl.WxOpenInMemoryConfigStorage;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * @author yl
+ */
+@Configuration
+@ConditionalOnProperty(
+  prefix = WxOpenProperties.PREFIX + ".config-storage", name = "type",
+  matchIfMissing = true, havingValue = "memory"
+)
+@RequiredArgsConstructor
+public class WxOpenInMemoryConfigStorageConfiguration {
+  private final WxOpenProperties properties;
+
+  @Bean
+  @ConditionalOnMissingBean(WxOpenConfigStorage.class)
+  public WxOpenConfigStorage wxOpenConfigStorage() {
+    WxOpenInMemoryConfigStorage config = new WxOpenInMemoryConfigStorage();
+
+    WxOpenProperties.ConfigStorage configStorageProperties = properties.getConfigStorage();
+    config.setWxOpenInfo(properties.getAppId(), properties.getSecret(), properties.getToken(), properties.getAesKey());
+    config.setHttpProxyHost(configStorageProperties.getHttpProxyHost());
+    config.setHttpProxyUsername(configStorageProperties.getHttpProxyUsername());
+    config.setHttpProxyPassword(configStorageProperties.getHttpProxyPassword());
+    if (configStorageProperties.getHttpProxyPort() != null) {
+      config.setHttpProxyPort(configStorageProperties.getHttpProxyPort());
+    }
+    return config;
+  }
+}
